# dist/public/swagger.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
...
            "RegistrationStatus": {
                "enum": [
                    "incomplete",
                    "complete",
                    "active"
                ],
                "type": "string"
            },
...
            "ISendOtpInput": {
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/OTPType"
                    }
                },
                "required": [
                    "email"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "IVerifyEmailInput": {
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "otp": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "otp"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "ICompleteBasicProfileInput": {
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "userName": {
                        "type": "string"
                    },
                    "referrer": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "firstName",
                    "lastName",
                    "password"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
            "Partial_Token_": {
                "properties": {},
                "type": "object",
                "description": "Make all properties in T optional"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/auth/send-otp": {
            "post": {
                "operationId": "SendOtp",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "hasActiveAccount": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "required": [
                                                "hasActiveAccount"
                                            ],
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ISendOtpInput"
                            }
                        }
                    }
                }
            }
        },
        "/auth/verify-email": {
            "post": {
                "operationId": "VerifyEmail",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "registrationStatus": {
                                                    "$ref": "#/components/schemas/RegistrationStatus"
                                                }
                                            },
                                            "required": [
                                                "registrationStatus"
                                            ],
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IVerifyEmailInput"
                            }
                        }
                    }
                }
            }
        },
        "/auth/complete-basic-profile": {
            "post": {
                "operationId": "CompleteBasicProfile",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ICompleteBasicProfileInput"
                            }
                        }
                    }
                }
            }
        },
        "/auth/complete-additional-profile": {
            "post": {
                "operationId": "CompleteAdditionalProfile",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ICompleteAdditionalProfileInput"
                            }
                        }
                    }
                }
            }
        },
        "/auth/setup-pin": {
            "post": {
                "operationId": "SetupPin",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Partial_Token_"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ISetupPinInput"
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "operationId": "Login",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Partial_Token_"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ILoginInput"
                            }
                        }
                    }
                }
            }
        },
        "/auth/confirm-pin": {
            "post": {
                "operationId": "ConfirmPin",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IConfirmPinInput"
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "operationId": "Logout",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": []
            }
        },
        "/auth/delete-account": {
            "delete": {
                "operationId": "DeleteAccount",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "message",
                                        "success"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IDeleteAccountInput"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "info": {
        "license": {
            "name": "ISC"
        },
        "contact": {},
        "title": "MD Exchange API",
        "description": "API for managing MD Exchange operations",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "localhost dev"
        }
    ]
}